requirements.txt
apscheduler==3.10.4
pandas==2.2.2
requests==2.32.3
pycryptodome==3.20.0
openpyxl==3.1.5